requirements.txt
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
coverage==7.6.1
Django==4.2.17
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
flake8==5.0.4
h11==0.14.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.9.1
pycparser==2.22
pyflakes==2.5.0
PyJWT==2.9.0
PySocks==1.7.1
pytest==8.3.4
pytest-cov==5.0.0
pytest-django==4.9.0
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.2
tomli==2.2.1
trio==0.27.0
trio-websocket==0.11.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
beautifulsoup4==4.12.3
black
flake8
